+++ iw/rejectanonymous/tests/test_rejectanonymousdocs.py
@@ -29,7 +29,15 @@ from Products.Five.testbrowser import Browser
 
 from base import TestCase
 
+import iw.rejectanonymous
+
 current_dir = os.path.dirname(__file__)
+
+def testSetUp(self):
+    # self is a base.TestCase
+    iw.rejectanonymous.valid_subparts = frozenset((
+        'portal_css', 'portal_javascripts', 'passwordreset'
+        ))
 
 def doc_suite(test_dir, setUp=None, tearDown=None, globs=None):
     """Returns a test suite, based on doctests found in /doctest."""
@@ -51,7 +59,7 @@ def doc_suite(test_dir, setUp=None, tearDown=None, globs=None):
         sys.path.append(package_dir)
 
     docs = []
-    for dir_ in ('doctests', 'docs'):
+    for dir_ in ('doctests',):
         doctest_dir = os.path.join(package_dir, dir_)
 
         # filtering files on extension
@@ -70,7 +78,7 @@ def doc_suite(test_dir, setUp=None, tearDown=None, globs=None):
 
 def test_suite():
     """returns the test suite"""
-    return doc_suite(current_dir)
+    return doc_suite(current_dir, setUp=testSetUp)
 
 if __name__ == '__main__':
     unittest.main(defaultTest='test_suite')